video-generation/requirements.txt
flask
flask-cors
pymongo
requests
pillow
imagehash
beautifulsoup4
selenium
scikit-image
opencv-python
werkzeug
fastapi
httpx
uvicorn
supabase
ffmpeg-python
torch

